# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Miaota Industrial Agent API
  description: |
    工业智能监控与诊断系统API
    
    ## 功能模块
    - **数据采集**: PLC数据采集与管理
    - **设备管理**: 设备状态监控与配置
    - **告警管理**: 告警规则与通知
    - **数据分析**: 异常检测与趋势分析
    - **知识库**: 故障诊断与知识检索
    - **报表**: 数据导出与可视化
    
  version: v1.0.0-beta1
  contact:
    name: Miaota Team
    email: team@miaota.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发环境
  - url: https://api.miaota.local/api/v1
    description: 生产环境
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /collection/start:
    post:
      tags: [数据采集]
      summary: 启动数据采集
      description: 启动指定设备的数据采集
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionStartRequest'
      responses:
        '200':
          description: 启动成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 采集已在运行
  /collection/stop:
    post:
      tags: [数据采集]
      summary: 停止数据采集
      parameters:
        - name: device_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 停止成功
...
  /data/query:
    post:
      tags: [数据采集]
      summary: 查询历史数据
      description: 查询指定时间范围的历史数据
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataQueryRequest'
      responses:
        '200':
          description: 查询结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataQueryResponse'
  /data/latest:
    get:
      tags: [数据采集]
      summary: 获取最新数据
      parameters:
        - name: tags
          in: query
          description: 要查询的点位标签
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: 最新数据

  # ============ 设备管理 ============
  /devices:
    get:
      tags: [设备管理]
      summary: 获取设备列表
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [s7, modbus]
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, error]
      responses:
        '200':
          description: 设备列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'

    post:
      tags: [设备管理]
      summary: 创建设备
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateRequest'
      responses:
        '201':
          description: 创建成功
...
  /analysis/anomaly:
    post:
      tags: [数据分析]
      summary: 异常检测分析
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnomalyAnalysisRequest'
      responses:
        '200':
          description: 分析结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyAnalysisResponse'
  /analysis/trend:
    get:
      tags: [数据分析]
      summary: 趋势分析
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            type: string
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 趋势数据
  /analysis/forecast:
    post:
      tags: [数据分析]
      summary: 预测分析
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag:
                  type: string
                horizon:
                  type: integer
                  description: 预测步数
                  default: 24
      responses:
        '200':
          description: 预测结果

  # ============ 知识库 ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证令牌
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API密钥认证
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [s7, modbus]
        host:
          type: string
        port:
          type: integer
        status:
          type: string
          enum: [online, offline, error]
        last_seen:
          type: string
          format: date-time
        tag_count:
          type: integer
    DeviceTag:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        data_type:
          type: string
        unit:
          type: string
        description:
          type: string
    DeviceCreateRequest:
      type: object
      required: [name, type, host, port]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [s7, modbus]
        host:
          type: string
        port:
          type: integer
        rack:
          type: integer
        slot:
          type: integer
        tags:
          type: array
          items:
            $ref: '#/components/schemas/DeviceTag'
...
    DataPoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
        quality:
          type: string
          enum: [good, bad, uncertain]
    DataQueryRequest:
      type: object
      required: [tags, start_time, end_time]
      properties:
        tags:
          type: array
          items:
            type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        aggregation:
          type: string
          enum: [raw, mean, sum, min, max]
          default: raw
        interval:
          type: string
          description: 聚合间隔（如 1m, 5m, 1h）
    DataQueryResponse:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/DataPoint'
    CollectionStartRequest:
      type: object
      properties:
        device_ids:
          type: array
          items:
            type: string
        scan_interval:
          type: integer
          default: 10

    # 分析相关
    AnomalyAnalysisRequest:
      type: object
      properties:
        tag:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        sensitivity:
          type: number
          default: 0.95
    AnomalyAnalysisResponse:
      type: object
      properties:
        anomalies:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
              value:
                type: number
              expected:
                type: number
              score:
                type: number
        total_points:
          type: integer
        anomaly_count:
          type: integer
        anomaly_rate:
          type: number

    # 知识库相关
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...